Move common function to formatter
