+++ internal/common/json.go
@@ -3,33 +3,3 @@
 // you may not use this file except in compliance with the Elastic License.
 
 package common
-
-import (
-	"bytes"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-)
-
-// JSONUnmarshalUsingNumber is a drop-in replacement for json.Unmarshal that
-// does not default to unmarshaling numeric values to float64 in order to
-// prevent low bit truncation of values greater than 1<<53.
-// See https://golang.org/cl/6202068 for details.
-func JSONUnmarshalUsingNumber(data []byte, v interface{}) error {
-	dec := json.NewDecoder(bytes.NewReader(data))
-	dec.UseNumber()
-	err := dec.Decode(v)
-	if err != nil {
-		if err == io.EOF {
-			return errors.New("unexpected end of JSON input")
-		}
-		return err
-	}
-	// Make sure there is no more data after the message
-	// to approximate json.Unmarshal's behaviour.
-	if dec.More() {
-		return fmt.Errorf("more data after top-level value")
-	}
-	return nil
-}
