+++ internal/formatter/json.go
@@ -7,7 +7,9 @@ package formatter
 import (
 	"bytes"
 	"encoding/json"
+	"errors"
 	"fmt"
+	"io"
 
 	"github.com/Masterminds/semver/v3"
 )
@@ -73,4 +75,26 @@ func (jsonFormatter) Encode(doc any) ([]byte, error) {
 
 	// Trimming to be consistent with MarshalIndent, that seems to trim the result.
 	return bytes.TrimSpace(formatted.Bytes()), nil
+}
+
+// JSONUnmarshalUsingNumber is a drop-in replacement for json.Unmarshal that
+// does not default to unmarshaling numeric values to float64 in order to
+// prevent low bit truncation of values greater than 1<<53.
+// See https://golang.org/cl/6202068 for details.
+func JSONUnmarshalUsingNumber(data []byte, v interface{}) error {
+	dec := json.NewDecoder(bytes.NewReader(data))
+	dec.UseNumber()
+	err := dec.Decode(v)
+	if err != nil {
+		if err == io.EOF {
+			return errors.New("unexpected end of JSON input")
+		}
+		return err
+	}
+	// Make sure there is no more data after the message
+	// to approximate json.Unmarshal's behaviour.
+	if dec.More() {
+		return fmt.Errorf("more data after top-level value")
+	}
+	return nil
 }
